Add build and tooling (#7)
* Rename README file

* Add chrome build

* Add tsconfigs

* Re-organize files

* Add path aliases

* Rewire imports to bundle with vite

* Update to use host_permissions

* Switch to webextension-polyfill package

* Update homepage

* Fix panel loading

The panel page and icon need to be referenced using an absolute path
that corresponds to their location in the build.
The panel.init function needed to be defined on window because we have
converted panel.js to a module.

* Update content script injection

Manifest V3 requires the scripting API when injecting scripts.

* Re-wire messaging to support service worker

Ports defined on the worker scope are not available when the service
worker is deactivated. We can use runtime messages from the content
script or create new ports from the devtools panel to wake the service
worker up.

* Remove counter example

We most likely will not need to send messages from the devtools panel to
the content script, so we can remove this example for now. If we do want
to add this messaging flow, we will need to implement service worker
support for it.

* Add Firefox build

* Revive background-devtools port on content message

We can re-establish the connection from the background script to the
devtools panel when the background script receives a message from the
content script.
The initial connection must be made from the panel because the panel may
not exist when the background script is inialized.

* Convert panel to a Svelte component

The Svelte component implements the UI and associated logic. We keep the
panel script for managing ports, commnunication, and evaluating scripts
in the inspected window.

* Remove uneeded web accessible resource declarations

* Set up eslint

* Add prettier config

The Svelte plugin uses prettier, so we need enforce rules here for the
formatting benefits.

* Lint all the things \o/

* Add placeholder TypeScript outDir

* Remove funding boilerplate

* Add Webnative icons

* Disable GitHub release action

We add and configure it when we are ready to release.

* Add linting to tests